requirements.txt
aiohttp~=3.7.4
python-dotenv~=0.15.0
requests~=2.25.1
scipy~=1.4.1
scikit-learn~=0.24.2
numpy~=1.18.0
pydantic~=1.8.2
pymystem3~=0.2.0
gensim~=3.8.3
uvicorn~=0.18.3
pathlib~=1.0.1
fastapi~=0.81.0
cyhunspell~=2.0.2
sentence-transformers~=2.2.2
pymssql~=2.2.2
pandas~=1.3.3
sklearn~=0.0
hunspell~=0.5.5
SQLAlchemy~=1.3.24
pytest~=7.1.2
pyodbc~=4.0.32
pyarrow~=3.0.0